😄🎉✨🎇🎆 Fini ! Algo fonctionne 🎊🎉🎈✨🎆🎇

+++ TP1-KNN/src/main/java/Iris.java
@@ -4,13 +4,13 @@ import java.util.Collections;
  * Created by Alexandre on 19/10/2016.
  */
 public class Iris {
-    float sepalLength;
-    float sepalWidth;
-    float petalLength;
-    float petalWidth;
+    double sepalLength;
+    double sepalWidth;
+    double petalLength;
+    double petalWidth;
     String type;
 
-    public Iris(float sL, float  sW, float  pL, float  pW, String type){
+    public Iris(double sL, double  sW, double  pL, double  pW, String type){
         this.sepalLength = sL;
         this.sepalWidth = sW;
         this.petalLength = pL;
